+++ packages/composites/augur/src/methods/resolveMarkets.ts
@@ -1,18 +1,12 @@
-import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig } from '@chainlink/types'
+import { Logger, Requester, Validator } from '@chainlink/ea-bootstrap'
+import { ExecuteWithConfig, Execute, AdapterResponse, AdapterRequest } from '@chainlink/types'
 import { Config } from '../config'
-import * as TheRundown from '@chainlink/therundown-adapter'
-import { ABI, Event, eventIdToNum, bytesMappingToHexStr } from './index'
+import { ABI, bytesMappingToHexStr, numToEventId } from './index'
 import { ethers } from 'ethers'
+import { theRundown, sportsdataio } from '../dataProviders'
 
 const resolveParams = {
-  sportId: true,
   contractAddress: true
-}
-
-const subDays = (date: Date, days: number): Date => {
-  date.setDate(date.getDate() - days)
-  return date
 }
 
 const eventStatus: { [key: string]: number } = {
@@ -22,73 +16,80 @@ const eventStatus: { [key: string]: number } = {
   'STATUS_CANCELED': 4
 }
 
+export interface ResolveEvent {
+  id: ethers.BigNumber
+  status: number
+  homeScore: number
+  awayScore: number
+}
+
+const statusCompleted = [
+  eventStatus['STATUS_CANCELED'],
+  eventStatus['STATUS_FINAL'],
+  eventStatus['STATUS_POSTPONED']
+  // TODO: What about other statuses in sportsdataio?
+]
+
+const tryGetEvent = async (dataProviders: Execute[], req: AdapterRequest): Promise<AdapterResponse> => {
+  let exception
+  for (let i = 0; i < dataProviders.length; i++) {
+    try {
+      return await dataProviders[i](req)
+    } catch (e) {
+      Logger.error(e)
+      exception = e
+    }
+  }
+  throw exception
+}
+
 export const execute: ExecuteWithConfig<Config> = async (input, config) => {
   const validator = new Validator(input, resolveParams)
   if (validator.error) throw validator.error
 
-  const sportId = validator.validated.data.sportId
   const contractAddress = validator.validated.data.contractAddress
-
   const contract = new ethers.Contract(contractAddress, ABI, config.wallet)
 
-  const params = { id: input.id, data: {
-      sportId,
-      date: new Date(),
-      endpoint: 'events'
-    }}
-  const theRundownExec = TheRundown.makeExecute()
-
-  const events = []
-  for (let i = 0; i < 2; i++) {
-    params.data.date = subDays(params.data.date, i)
-
-    const response = await theRundownExec(params)
-    events.push(...response.result as Event[])
+  const eventIDs: ethers.BigNumber[] = await contract.listResolvableEvents()
+  const events: ResolveEvent[] = []
+  for (const event of eventIDs) {
+    try {
+      const response = await tryGetEvent([theRundown.resolve, sportsdataio.resolve], {
+        id: input.id,
+        data: {
+          endpoint: 'event',
+          eventId: numToEventId(event)
+        }
+      })
+      events.push(response.result as ResolveEvent)
+    } catch (e) {
+      Logger.error(e)
+    }
   }
 
-  const statusCompleted = [
-    'STATUS_CANCELED',
-    'STATUS_FINAL',
-    'STATUS_POSTPONED'
-  ]
+  // Filters out events that aren't yet ready to resolve.
+  const eventReadyToResolve = events
+    .filter(({ status }) => statusCompleted.includes(status))
 
-  const filtered = events.filter(({ score: { event_status }}) => statusCompleted.includes(event_status))
-
-  const packed = filtered.map((event) => {
-    const status = eventStatus[event.score.event_status]
-    if (!status) return undefined
-
-    return packResolution(event.event_id, status, event.score.score_home, event.score.score_away)
-  }).filter((event) => !!event)
+  // Build the bytes32 arguments to resolve the events.
+  const packed = eventReadyToResolve.map(packResolution)
 
   let nonce = await config.wallet.getTransactionCount()
-  for (let i = 0; i < packed.length; i++) {
-    const eventId = eventIdToNum(filtered[i].event_id)
+  for (let i = 0; i < eventReadyToResolve.length; i++) {
+    Logger.info(`Augur: resolving event "${eventReadyToResolve[i].id}"`)
 
-    try {
-      const isResolved = await contract.isEventResolved(eventId)
-      if (isResolved) continue
-    } catch (e) {
-      // Skip if contract call fails, this is likely a
-      // market that wasn't created
-      continue
-    }
-
-    await contract.trustedResolveMarkets(packed[i], { nonce: nonce++ })
+    // This call both resolves markets and finalizes their resolution.
+    const tx = await contract.trustedResolveMarkets(packed[i], { nonce: nonce++ })
+    Logger.info(`Augur: Created tx: ${tx.hash}`)
   }
 
   return Requester.success(input.id, {})
 }
 
-export const packResolution = (
-  eventId: string,
-  eventStatus: number,
-  homeScore: number,
-  awayScore: number
-): string => {
+const packResolution = (event: ResolveEvent): string => {
   const encoded = ethers.utils.defaultAbiCoder.encode(
     ['uint128', 'uint8', 'uint16', 'uint16'],
-    [eventIdToNum(eventId), eventStatus, Math.round(homeScore*10), Math.round(awayScore*10)]
+    [event.id, event.status, Math.round(event.homeScore*10), Math.round(event.awayScore*10)]
   )
 
   const mapping = [16, 1, 2, 2]