+++ packages/core/legos/src/sources.ts
@@ -70,6 +70,7 @@ import * as reduce from '@chainlink/reduce-adapter'
 import * as renvm_adapter_address from '@chainlink/renvm-address-set-adapter'
 import * as satoshitango from '@chainlink/satoshitango-adapter'
 import * as sochain from '@chainlink/sochain-adapter'
+import * as sportsdataio from '@chainlink/sportsdataio-adapter'
 import * as stasis from '@chainlink/stasis-adapter'
 import * as taapi from '@chainlink/taapi-adapter'
 import * as therundown from '@chainlink/therundown-adapter'
@@ -155,6 +156,7 @@ export default {
   renvm_adapter_address,
   satoshitango,
   sochain,
+  sportsdataio,
   stasis,
   taapi,
   therundown,