+++ packages/sources/sportsdataio/src/adapter.ts
@@ -1,10 +1,10 @@
 import { Requester, Validator, AdapterError } from '@chainlink/ea-bootstrap'
 import { Config, ExecuteWithConfig, ExecuteFactory } from '@chainlink/types'
-import { makeConfig, DEFAULT_ENDPOINT } from './config'
-import { totalScore, events, event } from './endpoint'
+import { makeConfig, DEFAULT_SPORT } from './config'
+import { MMA, NFL } from './sport'
 
 const inputParams = {
-  endpoint: false,
+  sport: true,
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, config) => {
@@ -14,22 +14,19 @@ export const execute: ExecuteWithConfig<Config> = async (request, config) => {
   Requester.logConfig(config)
 
   const jobRunID = validator.validated.id
-  const endpoint = validator.validated.data.endpoint || DEFAULT_ENDPOINT
+  const sport = validator.validated.data.sport || DEFAULT_SPORT
 
-  switch (endpoint) {
-    case totalScore.NAME: {
-      return await totalScore.execute(request, config)
+  switch (sport.toLowerCase()) {
+    case MMA.NAME: {
+      return await MMA.execute(request, config)
     }
-    case events.NAME: {
-      return await events.execute(request, config)
-    }
-    case event.NAME: {
-      return await event.execute(request, config)
+    case NFL.NAME: {
+      return await NFL.execute(request, config)
     }
     default: {
       throw new AdapterError({
         jobRunID,
-        message: `Endpoint ${endpoint} not supported.`,
+        message: `Sport ${sport} not supported.`,
         statusCode: 400,
       })
     }