+++ packages/sources/therundown/src/endpoint/index.ts
@@ -1,2 +1,3 @@
 export * as totalScore from './totalScore'
 export * as events from './events'
+export * as event from './event'