+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New adapters:
   - `expert-car-broker` to get car prices
   - `binance` to get binance market data
+  - `sportsdataio` to get sports data from sportsdata.io
 
 ### Changed
 - Conflux adapter sending transaction optimization